working interrupt

+++ Snow-weAR/Core/Inc/GPS.h
@@ -12,12 +12,13 @@ typedef struct {
 	uint16_t	UTC_MicroSec;
 
 	float		Latitude;
-	double		LatitudeDecimal;
+	float		LatitudeDecimal;  // changed from double
 	char		NS_Indicator;
 	float		Longitude;
-	double		LongitudeDecimal;
+	float		LongitudeDecimal; // changed from double
 	char		EW_Indicator;
 
+	/*
 	uint8_t		PositionFixIndicator;
 	uint8_t		SatellitesUsed;
 	float		HDOP;
@@ -29,6 +30,7 @@ typedef struct {
 	uint16_t	AgeofDiffCorr;
 	char		DiffRefStationID[4];
 	char		CheckSum[2];
+	*/
 
 } GPGGA_t;
